Fix negative SPC volumes

+++ 3rdparty/snes-apu/src/dsp/dsp_helpers.rs
@@ -1,5 +1,5 @@
 pub fn multiply_volume(value: i32, volume: u8) -> i32 {
-    (value * (volume as i32)) >> 7
+    (value * (volume as i8 as i32)) >> 7
 }
 
 pub fn clamp(value: i32) -> i32 {
